+++ addon/system/file.js
@@ -1,19 +1,21 @@
 /* global moxie, plupload */
-import { copy } from 'ember-copy';
+import Ember from 'ember';
 
-import EmberArray, { A } from '@ember/array';
-import { alias, reads } from '@ember/object/computed';
-import RSVP from 'rsvp';
-import EmberObject, { get, computed } from '@ember/object';
+const get = Ember.get;
+const alias = Ember.computed.alias;
+const reads = Ember.computed.reads;
+
+const RSVP = Ember.RSVP;
 const moxieFileReader = function () {
   return new moxie.file.FileReader();
 };
 
 const keys = Object.keys;
+const computed = Ember.computed;
 
 const mergeDefaults = function (defaults, options) {
-  const unsetKeys = A(keys(defaults)).removeObjects(keys(options));
-  const settings = copy(options, true);
+  const unsetKeys = Ember.A(keys(defaults)).removeObjects(keys(options));
+  const settings = Ember.copy(options, true);
 
   for (let i = 0, len = unsetKeys.length; i < len; i++) {
     let key = unsetKeys[i];
@@ -39,8 +41,8 @@ const settingsToConfig = function (settings) {
   }, settings);
 
   if (headers.Accept == null) {
-    if (!EmberArray.detect(accepts)) {
-      accepts = A([accepts]).compact();
+    if (!Ember.Array.detect(accepts)) {
+      accepts = Ember.A([accepts]).compact();
     }
     headers.Accept = accepts.join(',');
   }
@@ -75,7 +77,7 @@ const settingsToConfig = function (settings) {
   @class File
   @extends Ember.Object
  */
-export default EmberObject.extend({
+export default Ember.Object.extend({
 
   /**
     The unique ID of the file.