+++ addon/system/flatten.js
@@ -1,10 +1,10 @@
-import { isArray } from '@ember/array';
+import Ember from 'ember';
 
 function flatten(array) {
   var flattened = [];
   for (let i = 0, len = array.length; i < len; i++) {
     let value = array[i];
-    if (isArray(value)) {
+    if (Ember.isArray(value)) {
       flattened.push(...flatten(value));
     } else {
       flattened.push(value);