+++ addon/system/sum-by.js
@@ -1,8 +1,9 @@
-import { A } from '@ember/array';
-import { get } from '@ember/object';
+import Ember from 'ember';
+
+var get = Ember.get;
 
 export default function (target, key) {
-  return A(target).reduce(function (E, obj) {
+  return Ember.A(target).reduce(function (E, obj) {
     return E + get(obj, key);
   }, 0);
 }